PokeGrids/requirements.txt
django
djangorestframework
django-filter
psycopg2
django-storages
gunicorn
